added libraries/tbl_columns_definition_form.lib.php . Added methods PMA_getFormsParameters, PMA_getHtmlForTableConfigurations, PMA_getHtmlForFooter, PMA_getHtmlForTableNameAndNoOfColumns, PMA_getHtmlForTableFieldDefinitions, PMA_getHtmlForTableCreateOrAddField
